requirements/base.txt
# Base
pytz==2018.7
python-slugify
Pillow==6.2.0
psycopg2 --no-binary psycopg2

# Django
django>=2.0.10

# Environment
django-environ>=0.4.5

# Django REST Framework
djangorestframework>=3.9.1
djangorestframework-simplejwt
django-filter>=2.0.0
django-cors-headers

# Django apps
django_taggit

# Passwords security
argon2-cffi>=18.3.0

# Static files
whitenoise>=5.0.0

# Celery
tornado
redis
django-redis
celery
flower
